# CW_AS/backend/utils/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Alumni Influencers API - w1956194",
    "version": "1.0.0",
    "description": "API documentation for the Alumni Influencers backend system. Student ID: w1956194"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
...
    "/api/analytics": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Main dashboard analytics data",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "x-api-key", "in": "header", "required": true, "schema": { "type": "string" } },
          { "name": "programme", "in": "query", "schema": { "type": "string" } },
          { "name": "graduationYear", "in": "query", "schema": { "type": "integer" } },
          { "name": "sector", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Analytics data" }
        }
      }
    },
    "/api/analytics/filters": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get filter options for analytics",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Filter options list" }
        }
      }
    }
  }
}